% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_data.R
\name{generate_data}
\alias{generate_data}
\title{Generate synthetic logistic regression data
Function taken from fastglm package}
\usage{
generate_data(N, k = 25, seed = 0)
}
\arguments{
\item{N}{Number of observations}

\item{k}{Number of predictors}

\item{seed}{seed for random number generation}
}
\value{
A list with matrix `X` and vector `y`
}
\description{
Generate synthetic logistic regression data
Function taken from fastglm package
}
